feat(frontend): db schema explorer + db aware AI
