+++ frontend/src/lib/components/codeGen/ScriptGen.svelte
@@ -9,11 +9,12 @@
 	import Popup from '../common/popup/Popup.svelte'
 	import { fade } from 'svelte/transition'
 	import { Icon } from 'svelte-awesome'
-	import { existsOpenaiResourcePath } from '$lib/stores'
+	import { dbSchema, existsOpenaiResourcePath } from '$lib/stores'
 	import type DiffEditor from '../DiffEditor.svelte'
 	import { scriptLangToEditorLang } from '$lib/scripts'
 	import type { Selection } from 'monaco-editor/esm/vs/editor/editor.api'
 	import type SimpleEditor from '../SimpleEditor.svelte'
+	import Tooltip from '../Tooltip.svelte'
 
 	// props
 	export let iconOnly: boolean = false
@@ -56,7 +57,8 @@
 			} else {
 				generatedCode = await generateScript({
 					language: lang,
-					description: funcDesc
+					description: funcDesc,
+					dbSchema: lang === 'postgresql' ? $dbSchema : undefined
 				})
 			}
 			funcDesc = ''
@@ -202,7 +204,6 @@
 			outerClasses="rounded shadow-xl bg-white border p-3 w-96"
 		>
 			<label class="block text-gray-900">
-				<div class="pb-1 text-sm text-gray-600">Prompt</div>
 				<div class="flex w-full">
 					<input
 						type="text"
@@ -230,6 +231,13 @@
 						<Icon data={faMagicWandSparkles} />
 					</Button>
 				</div>
+				{#if lang === 'postgresql' && $dbSchema}
+					<p class="text-sm mt-2"
+						>Will take into account the DB schema <Tooltip
+							>In order to better generate the script, we pass the selected DB schema to GPT-4.</Tooltip
+						></p
+					>
+				{/if}
 			</label>
 		</Popup>
 	{/if}
